% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_snps.R
\docType{data}
\name{manhattanFuncPtr}
\alias{manhattanFuncPtr}
\title{manhattanFuncPtr}
\format{
An object of class \code{XPtr} of length 1.
}
\usage{
manhattanFuncPtr
}
\description{
helper function to calculate distance
}
\keyword{datasets}
