% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.glmm-package.R
\docType{package}
\name{micro.glmm-package}
\alias{micro.glmm-package}
\alias{micro.glmm}
\title{The micro.glmm package}
\description{
micro.glmm is a library and R package for microbiome GLMM models
}
\keyword{glmm}
