% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{micro_glmm}
\alias{micro_glmm}
\title{micro_glmm}
\usage{
micro_glmm(
  obj.pop.strut,
  glm_fit0,
  GRM,
  copy_number_df,
  SPA = FALSE,
  scale_g = TRUE,
  log_g = TRUE
)
}
\arguments{
\item{obj.pop.strut}{output of pop_structure_test; GLMM of species with GRM accounted for}

\item{glm_fit0}{glm model. Model output with no sample relatedness accounted for}

\item{GRM}{Genetic Relatedness Matrix (from scripts or user) NxN matrix of sample relatedness}

\item{copy_number_df}{data frame with, gene_id, sample_name, and copy_number}

\item{SPA}{whether to run Saddle point approximation for pvalues (will slow down output)}

\item{scale_g}{whether to scale CNVs after log transform if using}

\item{log_g}{whether to log transform CNVs}
}
\value{
model output for the baseline structure glmm
}
\description{
Fit the CNV model with the random effects
}
