% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{micro_glmmSNP}
\alias{micro_glmmSNP}
\title{micro_glmmSNP}
\usage{
micro_glmmSNP(
  fit0,
  GRM,
  species_id,
  tau = c(0, 0),
  maxiter = 100,
  verbose = TRUE,
  tol = 1e-04
)
}
\arguments{
\item{fit0}{glm model. Logistic model output (with no sample relatedness accounted for)}

\item{GRM}{Genetic Realteness Matrix (from scripts or user) NxN matrix of sample realtedeness}

\item{tau}{vector for iniial values for the variance component parameter estimates}

\item{maxiter}{maximum iterations to fit the glmm model}

\item{verbose}{whether outputting messages in the process of model fitting}
}
\value{
model output for the baseline structure glmm
}
\description{
Fit the base model for SNP structure
}
